Add PS script to transform DICOM standard and download dependencies.
This commit adds a PowerShell script which includes two new functions: TransformDicomStandard and DownloadDeps. The former applies XSL transformations to XML data, and the latter downloads necessary dependencies if they're not present in the working directory. These changes are aimed at streamlining the process of setting up the environment and transforming the DICOM standard.